requirements.txt
python-dotenv
langchain-openai
langchain-chroma
langchain-community
langchain-core
chromadb